ci: fix fmt, clippy, and the per-file LOC cap
cargo fmt across the scraping and proxy-pool modules, drop a no-op f32
cast, and build GetCookiesParams by literal instead of default-then-assign.

Split switches_core_tests.rs (224 LOC) into core, version, and screen
test modules so every file stays under the 200-line cap.

+++ apps/engine/src/application/use_cases/stealth_config.rs
@@ -82,7 +82,7 @@ pub fn build_stealth_config(p: &Profile) -> StealthConfig {
         extensions_json: serde_json::to_string(&ext_entries).unwrap_or_default(),
         languages_json: serde_json::to_string(&p.locale.languages).unwrap_or_default(),
         hardware_concurrency: p.hardware.hardware_concurrency,
-        device_memory: p.hardware.device_memory_gb as f32,
+        device_memory: p.hardware.device_memory_gb,
     }
 }
